+++ crates/task-impls/src/request.rs
@@ -3,7 +3,7 @@ use std::{marker::PhantomData, sync::Arc, time::Duration};
 use async_broadcast::Sender;
 use async_compatibility_layer::art::{async_sleep, async_spawn, async_timeout};
 use async_lock::RwLock;
-use hotshot_task::task::TaskState;
+use hotshot_task::{broadcast_event, task::TaskState};
 use hotshot_types::{
     consensus::Consensus,
     message::{CommitteeConsensusMessage, DataMessage, Message, MessageKind, SequencingMessage},
@@ -20,10 +20,7 @@ use sha2::{Digest, Sha256};
 use tracing::{debug, error, info, instrument, warn};
 use vbs::{version::StaticVersionType, BinarySerializer, Serializer};
 
-use crate::{
-    events::{HotShotEvent, HotShotTaskCompleted},
-    helpers::broadcast_event,
-};
+use crate::events::{HotShotEvent, HotShotTaskCompleted};
 
 /// Amount of time to try for a request before timing out.
 const REQUEST_TIMEOUT: Duration = Duration::from_millis(500);
