+++ crates/task-impls/src/transactions.rs
@@ -10,7 +10,10 @@ use async_lock::RwLock;
 use hotshot_builder_api::block_info::{
     AvailableBlockData, AvailableBlockHeaderInput, AvailableBlockInfo,
 };
-use hotshot_task::task::{Task, TaskState};
+use hotshot_task::{
+    broadcast_event,
+    task::{Task, TaskState},
+};
 use hotshot_types::{
     consensus::Consensus,
     data::{null_block, Leaf},
@@ -32,7 +35,6 @@ use vbs::version::StaticVersionType;
 use crate::{
     builder::BuilderClient,
     events::{HotShotEvent, HotShotTaskCompleted},
-    helpers::broadcast_event,
 };
 
 /// Builder Provided Responses
