+++ crates/task/src/lib.rs
@@ -1,8 +1,41 @@
 //! Task primitives for `HotShot`
 
+use async_broadcast::{SendError, Sender};
+#[cfg(async_executor_impl = "async-std")]
+use async_std::task::JoinHandle;
+#[cfg(async_executor_impl = "tokio")]
+use tokio::task::JoinHandle;
+
 /// Simple Dependency types
 pub mod dependency;
 /// Task which can uses dependencies
 pub mod dependency_task;
 /// Basic task types
 pub mod task;
+
+/// Cancel a task
+pub async fn cancel_task<T>(task: JoinHandle<T>) {
+    #[cfg(async_executor_impl = "async-std")]
+    task.cancel().await;
+    #[cfg(async_executor_impl = "tokio")]
+    task.abort();
+}
+
+/// Helper function to send events and log errors
+pub async fn broadcast_event<E: Clone + std::fmt::Debug>(event: E, sender: &Sender<E>) {
+    match sender.broadcast_direct(event).await {
+        Ok(None) => (),
+        Ok(Some(overflowed)) => {
+            tracing::error!(
+                "Event sender queue overflow, Oldest event removed form queue: {:?}",
+                overflowed
+            );
+        }
+        Err(SendError(e)) => {
+            tracing::warn!(
+                "Event: {:?}\n Sending failed, event stream probably shutdown",
+                e
+            );
+        }
+    }
+}
